Finished the GIT STYLE GUIDE
This guide stablishes and standardizes the way we will be
working with git. It explains the different types of branches,
commit format, how to issue pull requests and raise issues.

This is the first version of this document so it can and should
be subject of a lot of changes as we better understand how to
work as a team.