Setup sphinx, with autodoc, apidoc and numpydoc
The initial setup for the creation and formatting of documentation
is completed. In the future the documentation for already
exisiting pieces of code such as (base and tosp_auth) should be
completed.

I suggest converting the testing environment into a testing and
documenting environment, in order to move those dependencies
away from all the environments.