Added Django CORS Functionality and Limited Users Token Creation
By default servers send a NOT ALLOW CROSS ORIGIN REQUEST to the client so that no request to another server can be sent. The offline client works as a local server that will load a page and then post data to the server to authenticate. Therefore we need a CROS ORIGIN REQUEST HEADER.

Also there was a discussion in an issue of limiting Tokens just for ADMINISTRACION and CAPTURISTA users.