Fix presentation of family members in focus mode
All family members ever created appeared in the focus mode, this has
been changed according to feedback from the stakeholders, to only
display the active family members. Along side those changes useless
code for a view that was never used in the final code has been removed
this view was the original implementation of the focus mode.